[DF-16098] Review fixes
Former-commit-id: dd31f4650d523b3e8e3129c425435e9f31428ab0

+++ Modules/TArc/Sources/TArc/Controls/PropertyPanel/Private/ReflectedPropertyItem.cpp
@@ -115,7 +115,7 @@ int32 ReflectedPropertyItem::LookupChildPosition(const std::shared_ptr<PropertyN
     DVASSERT(nodeParent != nullptr);
     while (true)
     {
-        if (positionCandidate >= children.size())
+        if (positionCandidate >= static_cast<int32>(children.size()))
         {
             break;
         }
